refactor(domain): split Domain.kt into 14 focused files
