+++ src/main/kotlin/io/sdkman/domain/AuditRepository.kt
@@ -1,0 +1,11 @@
+package io.sdkman.domain
+
+import arrow.core.Either
+
+interface AuditRepository {
+    suspend fun recordAudit(
+        username: String,
+        operation: AuditOperation,
+        data: Auditable,
+    ): Either<DatabaseFailure, Unit>
+}
