+++ src/main/kotlin/io/sdkman/domain/DatabaseFailure.kt
@@ -1,0 +1,16 @@
+package io.sdkman.domain
+
+sealed class DatabaseFailure(
+    override val message: String,
+    override val cause: Throwable,
+) : Throwable(message, cause) {
+    data class ConnectionFailure(
+        override val message: String,
+        override val cause: Throwable,
+    ) : DatabaseFailure(message, cause)
+
+    data class QueryExecutionFailure(
+        override val message: String,
+        override val cause: Throwable,
+    ) : DatabaseFailure(message, cause)
+}
