+++ src/main/kotlin/io/sdkman/domain/HealthCheck.kt
@@ -1,0 +1,8 @@
+package io.sdkman.domain
+
+data object HealthCheckSuccess
+
+enum class HealthStatus {
+    SUCCESS,
+    FAILURE,
+}
