+++ src/main/kotlin/io/sdkman/domain/HealthRepository.kt
@@ -1,0 +1,7 @@
+package io.sdkman.domain
+
+import arrow.core.Either
+
+interface HealthRepository {
+    suspend fun checkDatabaseConnection(): Either<DatabaseFailure, HealthCheckSuccess>
+}
