+++ src/main/kotlin/io/sdkman/domain/TagService.kt
@@ -1,0 +1,10 @@
+package io.sdkman.domain
+
+import arrow.core.Either
+
+interface TagService {
+    suspend fun deleteTag(
+        uniqueTag: UniqueTag,
+        username: String,
+    ): Either<DomainError, Unit>
+}
