+++ src/main/kotlin/io/sdkman/domain/VersionRepository.kt
@@ -1,0 +1,26 @@
+package io.sdkman.domain
+
+import arrow.core.Either
+import arrow.core.Option
+
+interface VersionRepository {
+    suspend fun read(
+        candidate: String,
+        platform: Option<Platform>,
+        distribution: Option<Distribution>,
+        visible: Option<Boolean>,
+    ): List<Version>
+
+    suspend fun read(
+        candidate: String,
+        version: String,
+        platform: Platform,
+        distribution: Option<Distribution>,
+    ): Option<Version>
+
+    suspend fun create(cv: Version): Either<DatabaseFailure, Int>
+
+    suspend fun findVersionId(uniqueVersion: UniqueVersion): Option<Int>
+
+    suspend fun delete(version: UniqueVersion): Int
+}
